Restructure directories and begin breaking out API from App
