+++ demo/index.html
@@ -30,47 +30,47 @@
         output including regular <code>console.log</code> calls will be displayed there. </p>
 
     <h2>API</h2>
-    <p>Optionally control the tool using JS via the API. The API is available on the global <code>FTwelve</code> object.
+    <p>Optionally control the tool using JS via the API. The API is available on the global <code>fTwelve</code> object.
     </p>
 
     <h3>Show</h3>
     <p>Attach F-Twelve to the DOM and display a collapsed tab at the bottom of the screen.</p>
-    <pre><code>FTwelve.show();</code></pre>
-    <button onclick="FTwelve.show()">Show</button>
+    <pre><code>fTwelve.show();</code></pre>
+    <button onclick="fTwelve.show()">Show</button>
 
     <h3>Hide</h3>
     <p>Remove F-Twelve from the DOM.</p>
-    <pre><code>FTwelve.hide();</code></pre>
-    <button onclick="FTwelve.hide()">Hide</button>
+    <pre><code>fTwelve.hide();</code></pre>
+    <button onclick="fTwelve.hide()">Hide</button>
 
     <h3>Disable</h3>
     <p>Completely detach and disable the tool including the keyboard shortcut and show method.</p>
-    <pre><code>FTwelve.disable();</code></pre>
-    <button onclick="FTwelve.disable()">Disable</button>
+    <pre><code>fTwelve.disable();</code></pre>
+    <button onclick="fTwelve.disable()">Disable</button>
 
     <h3>Enable</h3>
     <p>Enable the tool, this will also display it unless show is false.</p>
-    <pre><code>FTwelve.enable(show = true);</code></pre>
-    <button onclick="FTwelve.enable()">Enable</button>
-    <button onclick="FTwelve.enable(false)">Enable (Hidden)</button>
+    <pre><code>fTwelve.enable({show: true});</code></pre>
+    <button onclick="fTwelve.enable()">Enable</button>
+    <button onclick="fTwelve.enable({show: false})">Enable (Hidden)</button>
 
     <h3>On Show</h3>
     <p>Set a callback when showing the tool. Useful, for example, to use F-Twelve’s "debug hotkey" to enable additional
         "debug mode" features in the consuming application.</p>
-    <pre><code>FTwelve.onShow(() => {
+    <pre><code>fTwelve.onShow(() => {
 alert("Showing F-Twelve");
 });</code></pre>
-    <button onclick="FTwelve.onShow(function(){alert('Showing F-Twelve')})">Alert on show</button>
-    <button onclick="FTwelve.onShow(function(){})">Do nothing on show</button>
+    <button onclick="fTwelve.onShow(function(){alert('Showing F-Twelve')})">Alert on show</button>
+    <button onclick="fTwelve.onShow(function(){})">Do nothing on show</button>
 
     <h3>On Hide</h3>
     <p>Set a callback when removing from the DOM. Similarly as onShow, this can be used to disable the consuming
         application’s "debug mode" via keyboard shortcut.</p>
-    <pre><code>FTwelve.onHide(() => {
+    <pre><code>fTwelve.onHide(() => {
 alert("Hiding F-Twelve");
 });</code></pre>
-    <button onclick="FTwelve.onHide(function(){alert('Hiding f-Twelve')})">Alert on hide</button>
-    <button onclick="FTwelve.onHide(function(){})">Do nothing on hide</button>
+    <button onclick="fTwelve.onHide(function(){alert('Hiding f-Twelve')})">Alert on hide</button>
+    <button onclick="fTwelve.onHide(function(){})">Do nothing on hide</button>
 
     <script>
